requirements.txt
gunicorn
flask
Flask-APScheduler
flask-sqlalchemy
psycopg2-binary
requests2
beautifulsoup4
Flask-Mail
